update mongo uri
