+++ server.js
@@ -1,0 +1,25 @@
+require('dotenv').config()
+
+const express = require('express')
+const app = express()
+const cookieParser = require('cookie-parser')
+app.use(cookieParser())
+app.use(express.json())
+const port = (process.env.PORT || 4000)
+const http = require('http').createServer(app)
+const io = require('socket.io')(http)
+const MONGODB_URI = process.env.MONGODB_URI || 'mongodb://127.0.0.1/chatroom';
+
+const mongoose = require('mongoose')
+
+mongoose.connect(MONGODB_URI, {useNewUrlParser: true}, () => {
+    console.log('successfully connected to database')  
+})
+
+
+    const userRouter = require('./routes/User')
+    app.use('/user', userRouter)
+
+    http.listen(port, () => {
+        console.log(`Listening on port ${port}...`)
+    })
